cfehome/src/requirements.txt
dj-database-url==0.4.2
django-crispy-forms==1.6.1
gunicorn==19.7.1
olefile==0.44
Pillow==4.2.1
psycopg2==2.7.1
pytz==2017.2
